Replaced methods 'hasCauseAsAncestor' and 'hasExactCauseAsAncestor' in ThrowableAssert with the methods 'hasCauseOfType' and 'hasCauseOfExactType' in CauseHierarchyAssert. To get a CauseHierarchyAssert, users need to call the method 'causeHierarchy' in ThrowableAssert.
Replaced method 'hasCauseOfType' in ThrowableAssert with method 'cause', which returns another ThrowableAssert for the cause of the actual Throwable. Users can call 'isInstanceOf' in the returned ThrowableAssert.

Method 'isInstanceOf' throws IllegalArgumentException if the given class is null.        

Method 'isInstanceOfAny' throws IllegalArgumentException if the given class array is null or if any element in the class array is null.     

Improving test coverage.