Added method 'requireContent(Object)' to JTableCellFixture.
JTreeFixture now scrolls (up or down) to the row or path to select, drag or drop. Fixes issue 133.

Added method 'requireSelection(String)' to JTreeFixture. This method verifies that a JTree has the given path selected. The path is specified as a String, for example "root/node1/node1.1".

Added method 'requireSelection(int)' to JTreeFixture. This method verifies that a JTree has the row selected.

Added methods 'showPopupMenuAt(int)' and 'showPopupMenuAt(String)' to JTreeFixture. These methods shows a JPopupMenu in a given row and node path respectively.